opengl, render: fix room backgrounds always tiling to top left (incl. having draw_sprite_tiled use f64 instead of i32)

+++ src/render.rs
@@ -81,8 +81,8 @@ pub trait Renderer {
     fn draw_sprite_tiled(
         &mut self,
         texture: &AtlasRef,
-        x: i32,
-        y: i32,
+        x: f64,
+        y: f64,
         xscale: f64,
         yscale: f64,
         colour: i32,
